requirements.txt
tensorflow==2.9.1
beautifulsoup4==4.11.1
lxml==4.9.1
requests==2.28.1
xmltodict==0.13.0
pandas==1.5.2
htmlmin==0.1.12
matplotlib==3.6.2
seaborn==0.12.1
scikit-learn==1.1.3
geopy==2.3.0
python-telegram-bot==13.14
xgboost==1.7.1
numpy==1.23.5